requirements.txt
streamlit
langchain
faiss-cpu
openai
sqlite3
postgresql
langgraph
langsmith
pandas
numpy
dotenv